Refactor board animations into separate module for better code organization
